second attempt at mpSession; c++ library mimicking functionality currently offered by libctl
